+++ Malfunctions/State.cs
@@ -1,29 +1,28 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Malfunctions
+﻿namespace Malfunctions
 {
     internal class State
     {
-        // Whether the navigation malfunction is active.
-        public static bool malfunctionNavigation;
+        // Track our malfunctions.
+        public static Malfunction MalfunctionNavigation;
+        public static MalfunctionTeleporter MalfunctionTeleporter;
+        public static Malfunction MalfunctionDistortion;
+        public static MalfunctionPower MalfunctionPower;
 
-        // Store the state of sparks in the level.
-        public static UnityEngine.GameObject sparksNavigation;
+        public static void Load()
+        {
+            MalfunctionNavigation = new Malfunction();
+            MalfunctionTeleporter = new MalfunctionTeleporter();
+            MalfunctionDistortion = new Malfunction();
+            MalfunctionPower = new MalfunctionPower();
+        }
 
         public static void Reset()
         {
-            // Destroy any instantiated effect objects on reset.
-            if (sparksNavigation != null)
-            {
-                UnityEngine.Object.Destroy(sparksNavigation);
-            }
-
             // Reset all malfunctions to their default state.
-            malfunctionNavigation = false;
+            MalfunctionNavigation.Reset();
+            MalfunctionTeleporter.Reset();
+            MalfunctionDistortion.Reset();
+            MalfunctionPower.Reset();
         }
     }
 }